polowa okienek

+++ zlicz_bilon_rozdz5zad5.py
@@ -4,5 +4,5 @@ dwudziestoGroszowki = float(input("podaj ilość 20-groszówek: "))
 dziesiecioGroszowki = float(input("podaj ilość 10-groszówek: "))
 piecioGroszowki = float(input("podaj ilość 5-groszówek: "))
 
-lacznaWartoscBilonu = Decimal(float((piecdziesiecioGroszowki * 0.50) + (dwudziestoGroszowki * 0.20) + (dziesiecioGroszowki * 0.10) + (piecioGroszowki * 0.05)))
-print("łączna warość bilonu wynosi: ", round(lacznaWartoscBilonu, 2), "zł")
+lacznaWartoscBilonu = (float((piecdziesiecioGroszowki * 0.50) + (dwudziestoGroszowki * 0.20) + (dziesiecioGroszowki * 0.10) + (piecioGroszowki * 0.05)))
+print("łączna warość bilonu wynosi: ", round(lacznaWartoscBilonu,2), "zł")
